[!!!][TASK] Refactor eID handler for new dispatching API
The main() method is removed from the EidController and replaced
with a new tinyUrlRedirect() that accepts a request and reponse
object from the dispatcher.

The TinyUrlProcessorInclude is removed.

The Exception handling is adjusted. If no key is provided to
the eID controller we return a Bad Request Response.

The handling of the TinyUrlNotFoundException is simplified by
moving the Exception message into the class. The only constructor
parameter needed is the tiny URL key.