[!!!][TASK] Update tinyurl view helper, drop TYPO3 7.6 support
The view helper extends the standalone fluid AbstractViewHelper.

The view helper arguments are registered via the initializeArguments()
method.

We use the renderStatic() function instead of the render() function.

The handling for backwards compatible HTML escaping is dropped.

Update phpunit to version 7 and use static version of
nimut/testing-framework because it supports all TYPO3 versions now.

Resolves: #21