[!!!][TASK] Refactor TinyUrlGenerator
The deprecated method buildTinyUrl() is removed (use UrlUtils instead).

The getTinyUrl() method is deprecated and replaced by the
new generateTinyUrl() method.

Using the setters for the TinyUrl properties is deprecated. Instead
the TinyUrl model and repository should be used.

The settings class properties are replaced with a TinyUrl model
instance.

Constructor injection is used.