[FEATURE] Allow dictionaries with variable length an multibyte chars
The Base62UrlKeyGenerator is now an BaseXUrlKeyGenerator
depending on the length of the configured dictionary.

There will be no more Uninitialized string offset warnings when
using a dictionary with less than 62 chars.

By using mb_strlen() and mb_substr() is should now also be
possible to use a dictionary containing multiby characters.

Thanks to @thomashohn for reporting the issue and providing
the initial pull request.

Closes: #29