[!!!][FEATURE] Site configuration and customized base URLs
All extension configuration options can now be overwritten
in the Site configuration in the key 'tinyurls'. This
allows for example different storage pages for different Sites.

Also the base URL handling is now more flexible. A global
base URL can be configured in the Extension configuration,
it is possible to use the Site's "base" setting as base URL
or the base URL can be set in each Site config.

Additional Refactoring:

Centralize extension configuration data and config keys in new
ExtensionConfigurationData and ConfigKeys classes.

New interfaces are introduced (also for allow mocking readonly
classes): TinyUrlGeneratorInterface, UrlUtilsInterface

The TinyUrlDisplay class is now configured as a service.